Merge Spring re-organization into main branch. This should make it easier to make architectural changes.
Note that most code does not use Spring yet, but component scanning is ready and classes can add annotations as necessary.

Logging is now delegated to Spring, or Logback. Run with “—-debug” option or set in application.properties (or wherever) to see debug information.

Too many external dependencies were breaking when used as a module, so that code is now gone with this change. Look back at this change if you want to try to revert that. But it also means that the dependencies in the Gradle file are cleaner.

Signed-off-by: Lars Roe <lars@larsroe.com>